requirements.txt
fastapi
pandas
psycopg2-binary
python-multipart
uvicorn[standard]
